requirements.txt
torch
torchvision
tqdm
scikit-image
pillow
numpy
wandb
urllib
